Trying to improve spec by adding the concept of next generation
Cells lives inside a given population and a population must evolve into new generations.
In order to move to a new generation we must use population.next_generation and we should add cells to it.

A cell must to be placed into a specific point of the population, described by the x and y position.

In order to test it I've added a before_each macro to create the population in the beggining of each spec, but the
 following error was raised:

```
Error in line 2: while requiring "./spec/conway_game_of_life_spec.cr"

in spec/conway_game_of_life_spec.cr:5: undefined local variable or method 'before_each'

    before_each do
    ^~~~~~~~~~~
```

It happens that I was used to rspec's callback syntax and it does not works in default spec suite for crystal.