Adding specs releated to population class but without implementing their fixes
I've added the population_spec.cr to map Population class specs and I've learned a few things regarding to Crystal arrays and method overloads.

In Ruby you can create an array by:

```ruby
@cells = []
```

But, as Crystal is strongly typed we must declare the type of objects that are allowed in our array.
We can do that by:

```
@cells = Array(Type).new
```

It was a little hard to figure it out by my self but it makes a lot of sense.

Another thing I learned along the way was: if you does not provides a method override specific to nil values it will not allow passing nil values.

That's awesome. You don't need to add ifs to find out if a given parameter is nil. It just works in evaluation time.